refactor: extract DailyForecastItem parsing helpers
